AI/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
# cv2==4.5.5
Flask==2.1.3
Flask-Cors==3.0.10
flatbuffers==2.0
gast==0.5.3
gdown==4.5.1
gem==0.1.12
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpcio==1.43.0
h11==0.13.0
h5py==3.6.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.8.0
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.1.1
# numpy==1.21.4
oauthlib==3.2.0
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.15.0
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0